requirements.txt
streamlit
streamlit-option-menu
firebase-admin
google-cloud-secret-manager
google-cloud-aiplatform
google-cloud-logging